requirements.txt
attrs==19.3.0
Django==2.2.9
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.8.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
wcwidth==0.2.5
